feat(audio): 完成会议在线录音固定选路单声道阶段验证
- 保留双 MIC 采集，编码前固定选择 MIC0/MIC3，支持双声道回退
- 为 CHAT stream-only 会话启用单声道 Opus 编码
- 根据实际格式设置录音状态声道模式，修正 App 显示
- 完善实例资源与异常收尾，修复录音存储跨界余量丢失
- 增加主机行为验证，合并实施文档并保留实机日志

验证：固件构建和五组主机契约通过；实机 MIC0 编码
555 帧、22200 字节、failed=0，App 显示单声道。

边界：尚未接入动态择优；普通保存录音及 CALL 保持双声道。
App 音质、播放速度与时长、本地单声道回放仍待验收。

+++ SDK/audio/framework/nodes/sink_dev1_node.c
@@ -28,7 +28,7 @@ struct sink_dev1_hdl {
 
 extern int rdx_record_run_exit(void);
 extern int rdx_record_run_init(void);
-extern int rdx_record_run_data_handle(u8 *data, u16 len);
+extern int rdx_record_run_data_handle(u8 *data, u32 len);
 
 #ifdef BIT_RATE_TEST
 static int tx_speed_cal_timer = 0;
